Refactor DRB code into infrastructure
