+++ source/coordinator/GreenEnergyHub.Aggregation.Application/GreenEnergyHub.Aggregation.Application.csproj
@@ -20,7 +20,6 @@ limitations under the License.
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Azure.Databricks.Client" Version="1.1.2014.2" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
     <PackageReference Include="NodaTime.Serialization.SystemTextJson" Version="1.0.0" />
   </ItemGroup>
