+++ source/coordinator/GreenEnergyHub.Aggregation.Infrastructure/GreenEnergyHub.Aggregation.Infrastructure.csproj
@@ -26,6 +26,7 @@ limitations under the License.
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
+    <PackageReference Include="Microsoft.Azure.Databricks.Client" Version="1.1.2098.1" />
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.0.13" />
     <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="5.0.9" />
     <PackageReference Include="NodaTime.Serialization.SystemTextJson" Version="1.0.0" />
